requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.5
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
Faker==4.0.3
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
whitenoise==5.0.1
wincertstore==0.2
